# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Simple POS API
  version: 1.0.0
  description: |-
    Contract-first API for Simple POS (ADR-0008). Precedence: BUSINESS_RULES.md (BR v1.2) > this file > the UI/UX PDF > code.

    **Hosts**: API `https://api.cashier.hossam-ameen.online`, PWA `https://cashier.hossam-ameen.online` (separate origins,
    CORS limited to the PWA origin, ADR-0003).

    **Conventions** (ADR-0005, 0008, 0010, 0013)
    - Paths have no trailing slash. JSON is snake_case. Ids are integers. Timestamps are RFC 3339 UTC.
    - **Money is always an integer number of minor units (piastres) in a field ending in `_minor`** (BR-GEN-01). The server
      computes every total, change, expected cash and difference (BR-GEN-03).
    - Success body: `{success: true, message, data}`; lists: `data = {items, pagination, …}`; 204 has no body.
    - Error body: `{success: false, code, message, details}`; `code` is a BR §9 code with its BR §9 HTTP status and Arabic
      `message` (BR-GEN-06).
    - Auth: `Authorization: Bearer <access_token>` (15 min). The refresh token is an httpOnly cookie on the API host only,
      used by `/api/auth/refresh` and `/api/auth/logout` with `X-Requested-With: simple-pos` (ADR-0003, ADR-0011).
    - `x-roles` lists the roles allowed to call each operation (`PUBLIC`, `ANY_AUTHENTICATED`, `ADMIN`, `CASHIER`); every
      other authenticated role gets `FORBIDDEN_ROLE` (BR-ROLE-05). `x-br` lists the rules the operation enforces and
...

...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/categories:
    get:
      operationId: listCategories
      tags:
      - catalog
      summary: List categories (screen 08)
      description: |-
        All categories ordered by sort_order then name (BR-ITEM-06), with item counts. ADMIN only; cashiers use /api/catalog.

        BR: BR-ROLE-01, BR-ITEM-06.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-06
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      parameters:
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/CatalogStatus'
      responses:
        '200':
          description: Categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
    post:
      operationId: createCategory
      tags:
      - catalog
      summary: Create a category
      description: |-
        Name unique (`DUPLICATE_VALUE`); icon null or a key of the fixed set (BR-ITEM-07). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-07, BR §2 Category.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-07
      - BR §2 Category
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/categories/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getCategory
      tags:
      - catalog
      summary: Get a category
      description: |-
        ADMIN only.

        BR: BR-ROLE-01.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      responses:
        '200':
          description: The category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
    patch:
      operationId: updateCategory
      tags:
      - catalog
      summary: Update / enable / disable a category
      description: |-
        Disabling hides the category and all its items from ordering (BR-ITEM-05, BR-ITEM-01). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-05, BR-ITEM-07, AC-10.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-05
      - BR-ITEM-07
      - AC-10
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
    delete:
      operationId: deleteCategory
      tags:
      - catalog
      summary: Delete an empty category
      description: |-
        Only a category with zero items (any status) can be deleted; otherwise `CATEGORY_NOT_EMPTY` (BR-ITEM-05). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-05, BR-GEN-04, AC-10.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `CATEGORY_NOT_EMPTY`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-05
      - BR-GEN-04
      - AC-10
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - CATEGORY_NOT_EMPTY
      responses:
        '204':
          description: No content.
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
  /api/items:
    get:
      operationId: listItems
      tags:
      - catalog
      summary: List items (screen 09)
      description: |-
        Paginated, ordered by category sort_order, then item name. `search` matches the item name. ADMIN only.

        BR: BR-ROLE-01.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      parameters:
      - name: category_id
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/CatalogStatus'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A page of items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemPageEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
    post:
      operationId: createItem
      tags:
      - catalog
      summary: Create an item (screen 10)
      description: |-
        price_minor > 0 (BR-ITEM-02); name unique within its category (`DUPLICATE_VALUE`); description ≤ 500. Unknown category_id → `NOT_FOUND` (BR §9, US-23.8). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-02, BR §2 Item.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-02
      - BR §2 Item
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/items/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getItem
      tags:
      - catalog
      summary: Get an item
      description: |-
        ADMIN only.

        BR: BR-ROLE-01.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      responses:
        '200':
          description: The item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
    patch:
      operationId: updateItem
      tags:
      - catalog
      summary: Update / enable / disable an item (screens 09, 10)
      description: |-
        Price or name changes never affect existing order lines (BR-ITEM-03). A DISABLED item cannot be added to orders (BR-ITEM-01). Unknown path id or body `category_id` → `NOT_FOUND` (BR §9, US-23.8). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-01, BR-ITEM-02, BR-ITEM-03, AC-09, AC-10.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-01
      - BR-ITEM-02
      - BR-ITEM-03
      - AC-09
      - AC-10
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
    delete:
      operationId: deleteItem
      tags:
      - catalog
      summary: Delete a never-used item
      description: |-
        An item referenced by any order line → `ITEM_IN_USE` (disable it instead); a never-used item is hard-deleted (BR-ITEM-04, BR-GEN-04). ADMIN only.

        BR: BR-ROLE-01, BR-ITEM-04, BR-GEN-04, AC-10.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `ITEM_IN_USE`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-ITEM-04
      - BR-GEN-04
      - AC-10
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - ITEM_IN_USE
      responses:
        '204':
          description: No content.
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
  /api/catalog:
    get:
      operationId: getCatalog
      tags:
      - catalog
      summary: Orderable catalogue (screen 11)
      description: |-
        Only ACTIVE items in ACTIVE categories (BR-ITEM-01); categories by sort_order then name, items by name (BR-ITEM-06); each category carries its icon (BR-ITEM-07).

        BR: BR-ITEM-01, BR-ITEM-06, BR-ITEM-07.
        Errors: `UNAUTHENTICATED`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-ITEM-01
      - BR-ITEM-06
      - BR-ITEM-07
      x-error-codes:
      - UNAUTHENTICATED
      responses:
        '200':
          description: Catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: pos_refresh
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
        maxLength: 100
...
  responses:
    Error401:
      description: INVALID_CREDENTIALS (login) or UNAUTHENTICATED (missing/invalid/expired/revoked session).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: FORBIDDEN_ROLE, NOT_ORDER_OWNER, or USER_DISABLED (login).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: NOT_FOUND.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: State conflict — see the operation's x-error-codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: VALIDATION_ERROR, ORDER_EMPTY, DISCOUNT_INVALID or INSUFFICIENT_CASH.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    CatalogStatus:
      type: string
      enum:
      - ACTIVE
      - DISABLED
      description: Category/Item status.
    CategoryIcon:
      type: string
      description: Fixed built-in icon set (BR-ITEM-07, OQ-33). Keys are rendered by the PWA; no uploads. The 20 keys were
        approved by the owner at GATE B (D4).
      enum:
      - coffee
      - tea
      - juice
      - soft_drink
      - water
      - breakfast
      - sandwich
      - burger
      - pizza
      - grill
      - chicken
      - pasta
      - rice
      - salad
      - soup
      - appetizer
      - dessert
      - cake
      - ice_cream
      - bakery
...
    ErrorCode:
      type: string
      description: BR §9 codes (BR-GEN-06). The HTTP status of each code is fixed by BR §9. `INTERNAL_ERROR` (500) is technical,
        not a business rejection; approved by the owner at GATE B (D3) and added to BR §9 in BR v1.3 (ADR-0010).
      enum:
      - INVALID_CREDENTIALS
      - USER_DISABLED
      - TOO_MANY_ATTEMPTS
      - FORBIDDEN_ROLE
      - SHIFT_ALREADY_OPEN
      - NO_OPEN_SHIFT
      - SHIFT_HAS_OPEN_ORDERS
      - TABLE_OCCUPIED
      - TABLE_INACTIVE
      - ITEM_INACTIVE
      - ORDER_EMPTY
      - ORDER_NOT_EDITABLE
      - NOT_ORDER_OWNER
      - DISCOUNT_INVALID
      - INSUFFICIENT_CASH
      - ITEM_IN_USE
      - CATEGORY_NOT_EMPTY
      - USER_HAS_OPEN_SHIFT
      - LAST_ADMIN
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - NOT_FOUND
      - DUPLICATE_VALUE
      - IDEMPOTENCY_CONFLICT
      - INTERNAL_ERROR
    Pagination:
      type: object
      required:
      - page
      - page_size
      - total_items
      - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 1
    ErrorDetails:
      type: object
      description: Always present, may be empty. Known keys per code (ADR-0010).
      properties:
        fields:
          type: object
          description: 'VALIDATION_ERROR: field path → machine hints (English, never shown).'
          additionalProperties:
            type: array
            items:
              type: string
        open_order_numbers:
          type: array
          items:
            type: integer
          description: SHIFT_HAS_OPEN_ORDERS (BR-SHIFT-04).
        retry_after_seconds:
          type: integer
          nullable: true
          description: TOO_MANY_ATTEMPTS.
        item_ids:
          type: array
          items:
            type: integer
          description: ITEM_INACTIVE (BR-ITEM-01).
        field:
          type: string
          description: 'DUPLICATE_VALUE: the unique field (BR §9).'
      additionalProperties: true
    Error:
      type: object
      required:
      - success
      - code
      - message
      - details
      properties:
        success:
          type: boolean
          enum:
          - false
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Arabic message from BR §9, verbatim.
        details:
          $ref: '#/components/schemas/ErrorDetails'
...
    Category:
      type: object
      required:
      - id
      - name
      - sort_order
      - status
      - icon
      - item_count
      - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        sort_order:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/CatalogStatus'
        icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
        item_count:
          type: integer
          description: All items in the category, any status (screen 08).
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
    CategoryCreate:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: Unique (DUPLICATE_VALUE).
        sort_order:
          type: integer
          minimum: 0
          default: 0
        status:
          allOf:
          - $ref: '#/components/schemas/CatalogStatus'
          default: ACTIVE
        icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
          default: null
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: Unique (DUPLICATE_VALUE).
        sort_order:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/CatalogStatus'
        icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
    Item:
      type: object
      required:
      - id
      - category_id
      - category_name
      - category_icon
      - name
      - price_minor
      - description
      - status
      - in_use
      - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        category_id:
          type: integer
          format: int64
          minimum: 1
        category_name:
          type: string
        category_icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
          description: BR-ITEM-07.
        name:
          type: string
        price_minor:
          type: integer
          format: int64
          description: Price (> 0, BR-ITEM-02). Integer minor units (BR-GEN-01).
          minimum: 1
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/CatalogStatus'
        in_use:
          type: boolean
          description: Referenced by an order line → delete gives ITEM_IN_USE.
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
    ItemCreate:
      type: object
      required:
      - category_id
      - name
      - price_minor
      properties:
        category_id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
          description: Unique within its category.
        price_minor:
          type: integer
          format: int64
          description: Price > 0 (BR-ITEM-02). The PWA converts "45.50" → 4550. Integer minor units (BR-GEN-01).
          minimum: 1
        description:
          type: string
          maxLength: 500
          default: ''
        status:
          allOf:
          - $ref: '#/components/schemas/CatalogStatus'
          default: ACTIVE
    ItemUpdate:
      type: object
      properties:
        category_id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
          description: Unique within its category.
        price_minor:
          type: integer
          format: int64
          description: Price > 0 (BR-ITEM-02). Existing order lines keep their snapshot (BR-ITEM-03). Integer minor units
            (BR-GEN-01).
          minimum: 1
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/CatalogStatus'
    ItemPage:
      type: object
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CatalogItem:
      type: object
      required:
      - id
      - name
      - price_minor
      - description
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        price_minor:
          type: integer
          format: int64
          description: Current price. Integer minor units (BR-GEN-01).
          minimum: 1
        description:
          type: string
    CatalogCategory:
      type: object
      required:
      - id
      - name
      - sort_order
      - icon
      - items
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        sort_order:
          type: integer
        icon:
          allOf:
          - $ref: '#/components/schemas/CategoryIcon'
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItem'
    Catalog:
      type: object
      description: ACTIVE categories (sort_order, then name — BR-ITEM-06) with their ACTIVE items (BR-ITEM-01).
      required:
      - categories
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CatalogCategory'
...
    CategoryListEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    CategoryEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Category'
    ItemPageEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/ItemPage'
    ItemEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Item'
    CatalogEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Catalog'
...